fixed word wrapping issue
Added word-wrap property to make sure that no text overflows or gets clipped out. for example: some countries have so many time zones that the width of the container isn't enough to fit them, in such case we either increase the width which is more complicated to do per overflowing element, or doing it for all elements would break the design so making the text wraps to a new line is the easiest solution.